3d_dynamic_gnn/requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
pytorch-lightning>=2.0.0
h5py>=3.8.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
tensorboard>=2.12.0
pyyaml>=6.0
matplotlib>=3.6.0
seaborn>=0.12.0

